requirements.txt
pydantic>=2.7
rich>=13.7
typer>=0.12
pandas>=2.2
python-dotenv>=1.0
openai>=1.40
google-genai>=1.0
